requirements.txt
pip==24.0
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
streamlit==1.31.1
transformers==4.37.2
python-dotenv==1.0.1
pydantic-settings==2.2.0